Fix bug/improve lookuptable.c. Types guint->gint.

+++ interface.h
@@ -61,7 +61,7 @@ struct winctl {
 /* implemented by layout.c: */
 GtkWidget *createcplxplane(GtkWidget *txtcplx[4]);
 struct winctl *buildinterface(void);
-void setcplxplane(GtkWidget *txtcplx[4], const gdouble value[4], gdouble width, gdouble height);
+void setcplxplane(GtkWidget *txtcplx[4], const gdouble value[4], gint width, gint height);
 void restoredefaults(struct winctl *w);
 void statejctl(struct winctl *w);
 
@@ -70,7 +70,7 @@ void run_interface(gchar *file_name);
 void calc(GtkWidget *widget, struct winctl *w);
 void reset(GtkWidget *widget, struct winctl *w);
 void alloc_colors(struct iterate_param *it_param, struct winctl *w);
-void update_pixbuf(guint x, guint y, const IterationInfo *iterinfo,  const struct iterate_param *it_param);
+void update_pixbuf(gint x, gint y, const IterationInfo *iterinfo,  const struct iterate_param *it_param);
 void redraw_pixbuf(struct winctl *w);
 
 /* implemented by storedrawing.c */
@@ -78,7 +78,7 @@ void store_drawing_show(struct winctl *w);
 
 /* implemented by validate.c */
 gboolean validate_cplx(GtkWidget *txtcplx[4], gdouble cplx[4], GtkWidget *win);
-gboolean validate_itermax(GtkWidget *txtitermax, guint *itermax, GtkWidget *win);
+gboolean validate_itermax(GtkWidget *txtitermax, gint *itermax, GtkWidget *win);
 gboolean validate_degree(GtkWidget *txtdegree, gdouble *degree, GtkWidget *win);
 void errdialog(GtkWindow *win, gchar *msg);
